We changed all the things. This is the biggest commit ever
Solved the core spring loading problem.
Added the ClusterId, NodeID, and ConfigRoot as spring properties.
Added a class containing constants for the ReposeSpringProperties.
Removed ValvePowerFilter, because Power filter is now a
DelegatingFilterProxy to work with Spring
Telescoped the FilterContext loading stuff into one class with a rich
test suite
Created the CoreSpringProvider which uses TypesafeConfig to provide a
classpath scanning property that can be different in test and in prod.
Redesigned the SystemModelInterrogator to operate off of ClusterID and
NodeID instead of localhost interface and port list.
ServicePorts are basically not needed any longer at all.
ReposeInstance info is basically not needed at all.
Blew away the SpringConfiguration classes as they shouldn't be there.
The spring xml configs are still there for reference.
Moved FilterContextWrapper into it's own class, deleted the one that was
named the same that was used for building Valve. (will be done
    differently based on the prototype)
The core Unit tests all pass. (there are some ignored :( )