Throw the exceptions up to the FilterChain construction
THe filterContextFactory doesn't swallow exceptions any longer, and now
they have to be handled at the powerFilter. This will prevent the
filterchain from being replaced if a filter in the chain fails to
configure, load, or instantiate.

Prior behavior would be to skip that filter, and continue to do things
anyway, that seemed really bad.